+++ src/main/java/br/com/sudv/enfermeiros/crud/repository/EscalaRepository.java
@@ -1,7 +1,7 @@
 package br.com.sudv.enfermeiros.crud.repository;
 
-import br.com.sudv.enfermeiros.crud.model.Escala;
+import br.com.sudv.enfermeiros.crud.data.model.Escala;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EscalaRepository extends JpaRepository<Escala, Integer> {
+public interface EscalaRepository extends JpaRepository<Escala, Long> {
 }
